feat(agent): implement ExecutionSupervisor with DAG execution and review gate
Move complete DAG execution, review gate, and handoff logic from dispatch.ts to
ExecutionSupervisor. dispatch.ts is now a thin wrapper (118 lines).

Core features implemented in execution-supervisor.ts (1322 lines):
- DependencyGraph: buildDependencyGraph + assertAcyclicDependencyGraph
- Task scheduling: dispatchReadyTasks + startTaskRun + waitForNextResult
- ReviewGate: accept/reject with MAX_REJECTIONS_BEFORE_HANDOFF=3
- Handoff: rotateAgent + handoff document request
- Abort/timeout: executeChildRunWithAbort with signal handling

All 5 existing Dispatch tests pass. Type check clean.
